Move logger constant, remove fabric notice
