--- src/dummy/java/net/fabricmc/api/ModInitializer.java
@@ -1,7 +1,0 @@
-package net.fabricmc.api;
-
-
-
-public interface ModInitializer {
-	void onInitialize();
-}
