+++ src/main/java/at/yedel/yedelmod/launch/YedelModConstants.java
@@ -2,15 +2,9 @@ package at.yedel.yedelmod.launch;
 
 
 
-import org.apache.logging.log4j.LogManager;
-import org.apache.logging.log4j.Logger;
-
-
-
 public class YedelModConstants {
 	public static final String modid = "yedelmod";
 	public static final String name = "YedelMod";
 	public static final String version = "#version#";
-	public static final Logger yedelog = LogManager.getLogger("YedelMod");
 	public static final String logo = "§8§l- §9§lYedel§7§lMod §8§l-"; // "- YedelMod -"
 }
