--- src/main/resources/fabric.mod.json
@@ -1,19 +1,0 @@
-{
-  "contact": {
-    "homepage": "https://modrinth.com/mod/yedelmod",
-    "issues": "https://github.com/Yedelo/YedelMod/issues",
-    "sources": "https://github.com/Yedelo/YedelMod"
-  },
-  "description": "YedelMod only works on Forge 1.8.9!",
-  "entrypoints": {
-    "main": [
-      "at.yedel.yedelmod.launch.FabricNotice"
-    ]
-  },
-  "environment": "client",
-  "icon": "assets/yedelmod/yedelmod.png",
-  "id": "yedelmod",
-  "name": "YedelMod",
-  "schemaVersion": 1,
-  "version": "${version}"
-}
